Update chức năng cache dữ liệu xổ số

+++ springboot-postgresql-hibernate-crud/src/main/java/net/guides/springboot2/crud/repository/XosoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import net.guides.springboot2.crud.model.Xoso;
 
 @Repository
-public interface XosoRepository extends JpaRepository<Xoso, Integer>{
+public interface XosoRepository extends JpaRepository<Xoso, Long>{
 
 }
